requirements.txt
streamlit
numpy
pillow
matplotlib
requests
